requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
python-multipart
passlib
pyjwt
websockets
httpx